[Bronze II] Title: 숫자, Time: 264 ms, Memory: 23832 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -7,19 +7,19 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine(), " ");
 		
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
+		long A = Long.parseLong(st.nextToken());
+		long B = Long.parseLong(st.nextToken());
 		
         StringBuilder sb = new StringBuilder();
         
         if(B > A){
             sb.append(B - A - 1).append("\n");
-            for(int i = A + 1; i < B; i++)
+            for(long i = A + 1; i < B; i++)
                 sb.append(i).append(" ");
             System.out.println(sb.toString());
         } else if(A > B){
             sb.append(A - B - 1).append("\n");
-            for(int i = B + 1; i < A; i++)
+            for(long i = B + 1; i < A; i++)
                 sb.append(i).append(" ");
             System.out.println(sb.toString());
         } else {
